requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
reportlab>=4.0.0
Pillow>=10.0.0
psycopg2-binary>=2.9.0
